Update PDF.py
The generated file indeed appears to be in PDF format; however, it contains an extra period at the end that may catch the user's attention. For instance, if you name it "mystikal.pdf..app" it will display as "mystikal.pdf." which may seem unusual.

A more refined solution is available. When naming the file, utilize the \x20\x10 character sequence, as demonstrated below:
- a=$(echo -n "mystikal.pdf\x20\x10.app"); mv mystikal.app $a

This approach ensures that the file presents itself as "mystikal.pdf" presenting a more natural and expected appearance to the user.

Reference: https://www.youtube.com/watch?v=jMKtlS9H_TQ